Move break stringutil dependency on core
