+++ internal/stringutil/compare.go
