+++ internal/stringutil/convert.go
