+++ internal/stringutil/format.go
