+++ internal/testutil/baseline/error_baseline.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/microsoft/typescript-go/internal/ast"
 	"github.com/microsoft/typescript-go/internal/compiler"
-	"github.com/microsoft/typescript-go/internal/compiler/stringutil"
 	"github.com/microsoft/typescript-go/internal/core"
 	"github.com/microsoft/typescript-go/internal/tspath"
 	"gotest.tools/v3/assert"
@@ -179,7 +178,7 @@ func iterateErrorBaseline(t testing.TB, inputFiles []*TestFile, inputDiagnostics
 		markedErrorCount := 0
 		// For each line, emit the line followed by any error squiggles matching this line
 
-		lineStarts := stringutil.ComputeLineStarts(inputFile.content)
+		lineStarts := core.ComputeLineStarts(inputFile.content)
 		lines := lineDelimiter.Split(inputFile.content, -1)
 
 		for lineIndex, line := range lines {
